+++ CHANGES.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
-**UNRELEASED**
+**1.1.3**
 
 - ``catch()`` now raises a ``TypeError`` if passed an async exception handler instead of
   just giving a ``RuntimeWarning`` about the coroutine never being awaited. (#66, PR by